network/newstream: rename package, improve docs
